requirements/local.txt
-r base.txt # includes the base.txt requirements file

django-debug-toolbar==1.4
django-autofixture==0.11.0
